# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: microvm
  version: "1.0.0"
  description: |
    Every sandbox is a real virtual machine with its own kernel. Code inside one
    is assumed hostile.

    There are two ways to run code, and the difference is the point:

    - **Sandboxes** are held. You create one, run commands in it, and it is
      yours until you stop it or its TTL elapses. Creation fails when the node
      is full, so backpressure is yours to handle.
    - **Tasks** are queued. You hand work to the fleet and it runs when a slot
      frees anywhere. Submission never fails for capacity.

    Use a sandbox for several commands sharing state; use a task for throughput.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080/v1
    description: A local daemon. Bound to loopback by default, because this API creates VMs that run arbitrary code.
security:
  - bearerAuth: []
...
paths:
  /sandboxes:
    post:
      operationId: createSandbox
      summary: Create a sandbox
      description: |
        Boots a microVM and waits for it to answer. Returns once the sandbox is
        ready to run commands.

        Fails with `capacity_error` when the node has no room. That is
        deliberate: a sandbox is a reservation, so the caller is told
        immediately rather than left waiting. If you would rather wait, submit a
        task.
      tags: [Sandboxes]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SandboxCreateParams' }
      responses:
        '201':
          description: The sandbox is running and ready.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
        '500': { $ref: '#/components/responses/ApiError' }

    get:
      operationId: listSandboxes
      summary: List sandboxes
      description: |
        This node's sandboxes, newest first -- yours only. A sandbox belongs to the
        tenant that created it, and a list that reported every tenant's would hand
        out the IDs every other route is reached by. An admin key sees the whole
        node, because it is the operator's own.

        A stopped sandbox stays listed for a retention window and is then
        dropped, so this is not a history: a sandbox that ran an hour ago may no
        longer be here at all. Anything you need to keep -- the final metering
        above all -- comes from the reply to the delete, which is the only place
        it is ever stated.
      tags: [Sandboxes]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/EndingBefore'
        - name: state
          in: query
          description: Only sandboxes in this state.
          schema: { $ref: '#/components/schemas/SandboxState' }
        - name: tag
          in: query
          description: |
            Only sandboxes carrying this tag, written `key:value`. Repeat the
            parameter to require several: they AND, so `?tag=env:ci&tag=owner:me`
            returns the sandboxes carrying both. The split is at the first colon,
            so a value may contain colons and a key may not.

            The filter is node-local, because the list is: it searches the
            sandboxes this node holds, not the fleet's. On more than one node you
            get this node's answer, and finding a tagged sandbox anywhere means
            asking each node. A tag nobody set is an empty page, not an error --
            there is no registry of known keys to be wrong about.
          explode: true
          schema:
            type: array
            items: { type: string, pattern: '^[^:]+:.*$' }
          example: ["env:ci", "owner:billing-team"]
      responses:
        '200':
          description: A page of sandboxes, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SandboxList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    get:
      operationId: retrieveSandbox
      summary: Retrieve a sandbox
      description: |
        Returns the sandbox with live metering.

        A stopped sandbox is still here for a retention window, so a caller who
        polls learns *why* it stopped -- expired, idle, failed -- rather than
        being handed a bare 404 to interpret. Once the window elapses the record
        is dropped and this returns 404, which is why a 404 on an ID you know
        existed means "too late", not "never was".
      tags: [Sandboxes]
      responses:
        '200':
          description: The sandbox.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }

    delete:
      operationId: deleteSandbox
      summary: Delete a sandbox
      description: |
        Kills the VM and releases everything it held.

        Returns the sandbox with its final metering. Those numbers are sampled
        before the kill and cannot be obtained afterwards: the cgroup that holds
        them dies with the VM, so this reply is the only record of what the
        sandbox cost.
      tags: [Sandboxes]
      responses:
        '200':
          description: The sandbox, stopped, with its final cost.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/extend:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: extendSandbox
      summary: Extend a sandbox's TTL
      description: |
        Pushes the TTL deadline out to `ttl_seconds` from now, for the work that
        turned out longer than the caller guessed at create time. Without it the
        only way to buy time is to create a second sandbox, which does not have
        the first one's state.

        The new deadline is bounded by the host's maximum lifetime measured
        **from creation**, not from now. Extending therefore buys time and can
        never buy immortality: a caller that heartbeats forever still ends up
        with a sandbox that dies, which is the property that keeps a forgotten
        loop from pinning a slot for a week.

        Asking past that bound is a `400`, naming how many seconds are left,
        rather than a silent trim to the maximum. A caller told `200` for an hour
        they did not get plans for an hour and finds out when their work is
        killed halfway; an error arrives while they can still act on it. Read
        `expires` off the reply all the same -- that field is the deadline.

        It never brings a deadline forward. The deadline becomes the later of
        what it already was and `now + ttl_seconds`, because two callers
        heartbeating the same sandbox at different intervals would otherwise have
        the shorter one cut the longer one short. That also makes a retry
        harmless, which is why this route takes no Idempotency-Key.

        Extending a stopped sandbox is a `409`: there is no deadline left to
        move, and a caller heartbeating something already gone should learn it
        here rather than from the next exec.

        The idle timeout is untouched. A long TTL does not keep an idle sandbox
        alive; nothing except running something does.
      tags: [Sandboxes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SandboxExtendParams' }
      responses:
        '200':
          description: The sandbox, with its new `expires`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/executions:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: createExecution
      summary: Run a command
      description: |
        Starts a command and returns immediately with the execution. It does not
        wait for the command to finish.

        To follow the output, stream the execution; to collect it afterwards,
        retrieve it. Output is buffered on the host either way, so nothing is
        lost between starting the command and connecting to its stream.
      tags: [Executions]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecutionCreateParams' }
      responses:
        '201':
          description: The command is running.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }

    get:
      operationId: listExecutions
      summary: List a sandbox's executions
      tags: [Executions]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/EndingBefore'
      responses:
        '200':
          description: A page of executions, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/executions/{execution}:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
      - $ref: '#/components/parameters/ExecutionId'
    get:
      operationId: retrieveExecution
      summary: Retrieve an execution
      description: |
        Returns the execution and everything it printed.

        This works after the sandbox is gone. Output is buffered on the host
        precisely because the moment you most need it -- a timeout, a TTL, the
        OOM killer -- is the moment a guest-side buffer would die with the VM.
      tags: [Executions]
      responses:
        '200':
          description: The execution.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/executions/{execution}/stream:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
      - $ref: '#/components/parameters/ExecutionId'
    get:
      operationId: streamExecution
      summary: Stream an execution's output
      description: |
        Server-sent events, one per output chunk, as the process produces them.
        The stream ends with an `exit` or `error` event.

        The stream replays from the start of the execution, so connecting late
        loses nothing and a dropped connection can simply reconnect. This is why
        starting a command and streaming it are separate calls: a single call
        that did both would make the output unrecoverable the moment the
        connection broke.

        SSE rather than WebSockets because the traffic is one-directional: there
        is no upgrade, no ping/pong, and it survives any proxy that speaks HTTP.
      tags: [Executions]
      responses:
        '200':
          description: A stream of frames.
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/Frame' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/executions/{execution}/cancel:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
      - $ref: '#/components/parameters/ExecutionId'
    post:
      operationId: cancelExecution
      summary: Cancel an execution
      description: |
        Signals the process. The signal reaches the whole process group, so a
        program that spawned children does not leave them running.

        Defaults to SIGKILL. Hostile code is not owed a chance to clean up, and
        a graceful signal it can ignore is a way to outlive its own cancel.
      tags: [Executions]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecutionCancelParams' }
      responses:
        '200':
          description: The execution.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/files:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: createFile
      summary: Upload a file
      description: Writes a file into the sandbox, creating parent directories.
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileCreateParams' }
      responses:
        '201':
          description: The file was written.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/File' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }

    get:
      operationId: retrieveFile
      summary: Download a file
      description: |
        The path is a query parameter rather than part of the route because a
        file path contains slashes, and encoding them into a path segment would
        make every caller escape their own paths to name a file.
      tags: [Files]
      parameters:
        - name: path
          in: query
          required: true
          description: Absolute path inside the sandbox.
          schema: { type: string, example: "/app/out.json" }
      responses:
        '200':
          description: The file's contents.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/files/batch:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: createFiles
      summary: Upload several files
      description: |
        Writes a set of files in one request, in the order given, creating parent
        directories.

        Each path is named at most once. Two entries for one path would report two
        sizes for one file and leave only the later on disk, which makes the reply
        a statement about nothing, so the batch is refused rather than resolved
        last-wins.

        One round trip instead of N. A project is rarely a single file, and
        twenty uploads spend twenty HTTP requests and twenty vsock round trips on
        work the guest can do in one.

        Validation is all-or-nothing: every entry is checked -- path, mode, size
        -- before a single byte enters the sandbox, so a batch with one bad path
        writes nothing at all rather than half a project. Writing is not
        transactional and cannot be: once the third file is on the guest's disk
        there is no unwriting it. A write that fails partway therefore fails the
        request with the files before it already written, and the order is the
        order you gave precisely so a caller can say which those were.

        The batch is one request body, so it is bound by the daemon's body cap
        for all of the files together -- and base64 makes the payload a third
        larger than the bytes it carries.
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileBatchCreateParams' }
      responses:
        '201':
          description: |
            Every file was written, in the order given. A list envelope for
            consistency with every other collection, not because it pages:
            `has_more` is always false.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileList' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /sandboxes/{sandbox}/dirs:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: createDir
      summary: Create a directory
      description: |
        Creates a directory and any parents it needs.

        Uploading a file already creates its parents, so this route is for the
        directory that stays empty: an output directory a command writes into, a
        mount point, a layout a build tool expects to exist before it starts.

        Existing is success, not conflict -- `mkdir -p` semantics are what a
        caller building a tree wants, and they make a retry safe. An existing
        *file* at the path is a conflict, because that is a caller confusing two
        different things rather than repeating themselves.

        The directory is created `0755`, and so is every parent. There is no
        `mode` knob: a directory without its execute bit cannot be entered, so
        most of what one could ask for here is a way to create something
        unusable.
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DirectoryCreateParams' }
      responses:
        '201':
          description: The directory exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Directory' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /tasks:
    post:
      operationId: createTask
      summary: Submit a task
      description: |
        Queues work for the fleet. Unlike creating a sandbox, this never fails
        for capacity: the task waits for a slot on any node.

        The queue is the source of truth. No node is told what to run; each
        pulls when it has a free slot, which is why adding the three-hundredth
        node needs no coordination and losing one needs no detection.
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreateParams' }
      responses:
        '202':
          description: The task is queued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
        '503': { $ref: '#/components/responses/QueueUnreachable' }
  /tasks/{task}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      operationId: retrieveTask
      summary: Retrieve a task
      tags: [Tasks]
      responses:
        '200':
          description: The task, with its result once it has one.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
        '503': { $ref: '#/components/responses/QueueUnreachable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Tokens are compared in constant time.
...
  responses:
    InvalidRequest:
      description: The request is malformed or contradicts itself.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: The token is missing or wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: The object already exists, or is in a state that forbids this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ApiError:
      description: Something failed on our side.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...